Create galaxy samples with custom filter
An alternative to opencosmo package, tested to access diffsky catalogs on
NERSC. I wrote this test to independently validate the galaxy samples obtained
from opencosmo.

The access to all the diffsky catalogs at NERSC is set by the path given in the
code. One can edit the path in the code and apply to any other catalog in
principle.

Features:
* Modify the 'cuts' function to include any additional cuts on the galaxy sample.
* Save the filtered catalog in hdf5 format.